requirements.txt
paramiko>=3.3.1
pymongo>=4.6.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
bcrypt>=4.1.2
PyJWT>=2.8.0
python-multipart>=0.0.9
Jinja2>=3.1.3
